Add filter to event callbacks
