+++ lib/msal-browser/src/controllers/IController.ts
@@ -26,6 +26,7 @@ import { AuthenticationResult } from "../response/AuthenticationResult.js";
 import { EventCallbackFunction } from "../event/EventMessage.js";
 import { ClearCacheRequest } from "../request/ClearCacheRequest.js";
 import { InitializeApplicationRequest } from "../request/InitializeApplicationRequest.js";
+import { EventType } from "../event/EventType.js";
 
 export interface IController {
     // TODO: Make request mandatory in the next major version?
@@ -49,7 +50,7 @@ export interface IController {
         accountId?: string
     ): Promise<AuthenticationResult>;
 
-    addEventCallback(callback: EventCallbackFunction): string | null;
+    addEventCallback(callback: EventCallbackFunction, eventTypes?: Array<EventType>): string | null;
 
     removeEventCallback(callbackId: string): void;
