+++ lib/msal-browser/src/index.ts
@@ -86,6 +86,7 @@ export {
     PopupEvent,
 } from "./event/EventMessage.js";
 export { EventType } from "./event/EventType.js";
+export { EventHandler } from "./event/EventHandler.js";
 
 export {
     SignedHttpRequest,
